+++ nosr.c
@@ -210,7 +210,7 @@ static int search_metafile(const char *repo, struct pkg_t *pkg,
 			continue;
 		}
 
-		if(!found && config.filterfunc(&config.filter, buf.line, len, config.icase) == 0) {
+		if(!found && config.filterfunc(&config.filter, buf.line, (int)len, config.icase) == 0) {
 			if(config.verbose) {
 				if(asprintf(&line, "%s/%s %s\t/%s", repo, pkg->name, pkg->version, buf.line) == -1) {
 					fprintf(stderr, "error: failed to allocate memory\n");
@@ -236,7 +236,7 @@ static int list_metafile(const char *repo, struct pkg_t *pkg,
 	const char * const files = "%FILES%";
 	struct archive_read_buffer buf;
 
-	if(config.filterfunc(&config.filter, pkg->name, (size_t)-1, config.icase) != 0) {
+	if(config.filterfunc(&config.filter, pkg->name, -1, config.icase) != 0) {
 		return 1;
 	}
